requirements.txt
fastapi
uvicorn
jinja2
pydantic
pytest
httpx
python-multipart
passlib[bcrypt]
python-jose